+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 ## 3.74.0 (Unreleased)
+
+### Added
+- Support for update `license_model` in `oci_database_db_system`
+- Support for ADB-S Version Upgrade 19c (manual)
+- Support restore feature for key and vault
+
+### Fixed
+- Fix `streaming_stream_pool_resource` send empty key
+
 ## 3.73.0 (April 29, 2020)
 
 ### Added